feat: Implement nested volume model for LUKS and LVM support
- Added a new `VolumeNode` structure to represent nested volumes in `disks-dbus`, allowing for a tree-like representation of block devices.
- Enhanced the `DriveModel` to include a list of volumes, supporting both physical and logical volumes.
- Implemented LUKS unlock/lock operations using the UDisks2 `Encrypted` interface, exposing these actions in the UI.
- Added functionality to enumerate contained filesystems and render them as nested child rows in the UI.
- Integrated best-effort LVM support, allowing for the enumeration of logical volumes (LVs) and mapping their device paths back to UDisks block objects.
- Updated the UI to prompt for a passphrase when unlocking encrypted volumes and to display the locked/unlocked state of containers.
- Ensured that the existing partition segmentation UI remains compatible with the new volume representation.
- Added comprehensive tests and documentation for the new features.

+++ disks-dbus/src/usage.rs
@@ -24,14 +24,14 @@ pub fn usage_for_mount_point(mount_point: &str, filesystem: Option<&str>) -> Res
 
     let stat = unsafe { stat.assume_init() };
     let frsize = if stat.f_frsize > 0 {
-        u64::from(stat.f_frsize)
+        stat.f_frsize
     } else {
-        u64::from(stat.f_bsize)
+        stat.f_bsize
     };
 
-    let total = u64::from(stat.f_blocks).saturating_mul(frsize);
-    let free = u64::from(stat.f_bfree).saturating_mul(frsize);
-    let available = u64::from(stat.f_bavail).saturating_mul(frsize);
+    let total = stat.f_blocks.saturating_mul(frsize);
+    let free = stat.f_bfree.saturating_mul(frsize);
+    let available = stat.f_bavail.saturating_mul(frsize);
     let used = total.saturating_sub(free);
     let percent = if total == 0 {
         0
